Introduce some bugs to test detection
** Why are these changes being introduced:

* It would be helpful to have a few bugs somewhere to demonstrate how
  CodeClimate and GitHub Actions behave now.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/lm-10

** How does this address that need:

This introduces some bad code (of varying degrees) into the footer
template, so that both CodeClimate and GitHub Actions have something to
do:

* There are aesthetic violations, like comments being formatted poorly.
* There are naming convention problems, as WordPress does not like
  variable names like $thisFoo.
* There are some security problems, because blindly using `echo` to
  write output without escaping it can be problematic.

CodeClimate should flag all of these issues, but those violations can
be dismissed within CodeClimate by the reviewer. Future PRs will not
mention the issue unless those lines of code get changed.

GitHub Actions should flag the security problem on line 16 for every
future PR. However, the same problem on line 19 should not be noted (by
either tool) because the comment on line 18 causes it to be ignored.
This sort of comment-based whitelisting is the only way I know of to
make GitHub Actions overlook a problem.

Although I think it is possible to merge a PR while these objections are
present, but it obviously should not be our practice to do so.

** Document any side effects to this change:

* This commit will obviously not be present when this merges.