Cleanup news theme javascript
** Why are these changes being introduced:

* The javascript files used by the news theme need some cleanup - not
  every file in the theme appears to be loaded, and there is some mixing
  of third-party files and code written internally.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/lm-61

** How does this address that need:

This tries to streamline the javascript files within the theme repo:

* The unused scrollPosts.js file is removed
* The third-party lazyload library is moved from js/ to js/libs, and
  named lazyload.min.js to reflect its minification.
* The functions handling registration and enqueuing of javascript are
  merged with the one handling stylesheets - which brings this theme
  into a similar structure as the other themes (a single function to
  handle all styles and javascript files).

Please note, however, that the news theme also has inline javascript in
many of its page templates - and those templates have not been moved
into this theme yet. There is a coming ticket to import these page
templates, at which point we will have an option to leave that JS as
inline or split off to separate files.

** Document any side effects to this change:

* None, although as noted above there is additional javascript work
  coming in a future ticket, because of inline js in page templates.