Implements location and location-2021 templates
The tabbed display on the classic location template is still not working - will need a future ticket.

This brings along the Location field group, now that we have these two templates in the theme.

It also brings in the cf() wrapper function in the theme, which is a wrapper for get_post_meta.