Whitelist navwalker library
The third-party navwalker library is causing problems
with our code analysis tools, noting problems that are
not in our capability to fix without forking the project.

This change starts ignoring the navwalker file from our
analysis tools. There is a separate ticket in our backlog
to extract this to a separate dependency - LM-162