Document how to create / modify custom post types
** Why are these changes being introduced:

* The process we are using for managing custom post types - and the
  fields within them - has been assembled from a few plugins and the
  WordPress documentation. That process needs to be documented clearly
  so that others (including ourselves in the future) can use it
  successfully.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/lm-146

** How does this address that need:

This attempts to write a "how-to" style set of instructions that we
can use in the future to continue using the processes set up for our
custom post types. It attempts to avoid more general descriptions of our
setup.

Tasks documented are:
1. How to create a new post type
2. How to modify an existing post type (and its implications)

** Document any side effects to this change:

* This text would not have fit well within our existing project readme,
so I'm using this ticket to set up a shell of the Diátaxis framework.

There is a minimal readme with a docs/ folder with a link to Diátaxis,
and folders with .gitkeep files for the other three categories of
documentation.