Updates 404 page content
Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/LM-270

How does this address that need:

* Updates markup to match requested changes in linked ticket

Document any side effects to this change:

The 404 template and the default search results template both share the
same search form partial. We don't generally expect the default search
template to be used, but the changes to 404 do cause that to update as
well.