Refactor analytics plugin to use Matomo instead of GA
Why these changes are being introduced:

We need to migrate off of Google Analytics by July 1, and the
Pantheon migration is a good opportunity to move to Matomo
on libraries.mit.edu.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/UXWS-1555

How this addresses that need:

This replaces GA settings and tracking code in the analytics
plugin with Matomo settings and tracking code.

Side effects of this change:

* GA tracking code has also been removed from the multisearch
widget and libraries theme. This code was part of the 'discovery'
GA view that has not been used for several years.
* When we merge this, analytics in Pantheon will be reported
to Matomo only. This should not result in any data loss, as the
leagcy (current prod) site will continue reporting to GA until we
launch the migrated site. UXWS stakeholders are aware of this
decision and approve of it.