Remove "no photo" image from child theme
The MOH site had a feature where we would load the Libraries' logo
if an interviewee had no headshot in our system. The logo we were
using was out of date, and we've decided we do not want this feature
anyway.

This removes that default use of an image when no headshot is defined.