Restructure page headings on singler interview template
The heading structure used by the Interview template had some poor
choices, with a poorly placed h2 and the primary heading being marked
as an h3 (which is a peer to some internal headings farther down).

This rationalizes the heading structure a bit, but does not touch the much
larger problem of using the organization name as the h1 - which still
needs to be changed, but this is not the ticket for that work.