Add dashicons to front end of child theme
The Dashicon styles are automatically enqueued for use in the admin
interface, but are not loaded by default for public-facing pages. We use
one of these icons on the interview page, however.

This change makes sure that those dashicons are enqueued on the
front end as well, by declaring them a dependency of the main child
stylesheet.