Add conditional config for the UX tier
** Why are these changes being introduced:

We have a long-term multidev for the UX team to use for content work,
but it isn't consistently implemented between several names. This causes
problems when trying to access the network dashboard and password reset
interfaces, because the inconsistent names provoke a server error.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/pw-92

** How does this address that need:

This adds a conditional to web/wp-config.php to standardize how the UX
tier is named. This is the same conditional where we assign custom names
for the Dev, Test, and Live tiers - so this is a consistent place to
assign a custom name for a multidev.

** Document any side effects to this change:

None that I can see - although this is the first time we're using this
approach for a multidev, so that may be worth noting.