Skip Sentry configuration in Lando environments
** Why are these changes being introduced:

We integrate with Sentry provided a given ENV value is defined in
secrets.json, but this can result in local environments sending
exceptions to Sentry - which be quite noisy and undesired.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/pw-106

** How does this address that need:

This adds an additional check around the Sentry configuration block,
making sure that the PANTHEON_ENVIRONMENT variable is not sent to
"lando" when integrating with Sentry.

** Document any side effects to this change:

None