** Why are these changes being introduced: Our current Alerts in Wordpress don't follow our style guide, and aren't consistent inside of Wordpress. The particular example in PW-108 is handcoded and uses inline style.
We also now have a new Alert pattern styling as part of our new design system efforts.
*

** Relevant ticket(s):
PW-108 - https://mitlibraries.atlassian.net/browse/PW-108

* https://mitlibraries.atlassian.net/browse/

** How does this address that need:
This work implements the new alert styling inside of our Wordpress parent theme. The new class is prefixed with "mitlib-" to avoid conflict with styles from other plugins or themes.

This particular instance uses the default informational style, but classes have been added to upgrade this to danger, warning, or success in future use cases.
*

** Document any side effects to this change:
This work uses icons that already existed in FontAwesome. We'll need to upgrade to a new version to use the ideal icons. The icon choices are semantically close and will be fine until we're able to upgrade.
*