Update alerts to use new FA icons and update alert on location pages
** Relevant ticket(s):
https://mitlibraries.atlassian.net/browse/PW-113

** How does this address that need:
This commit adds the newest version of FontAwesome via our CDN, and changes the icons appropriately for the new alert styles.

** Document any side effects to this change:
This commit breaks some of the old references to FontAwesome, which will need to be addressed in the following commits.

Fixed icons in site-wide alerts, and remapped styles for exclamation icon to fit