Add filter to allow conditional max cache ages
** Why are these changes being introduced:

There are several paths which have proven problematic under our current
caching setup (where content lasts for a week before re-rendering)

1. The hours grid, which accepts a URL parameter to highligh specific
   days.
2. The news site's events page
3. The exhibit site's homepage
4. The exhibit site's event index, which contains lists of current,
   future, and past exhibits.

Each of these pages doesn't work well when cached for more than a day,
because content can get out of date after 24 hours.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/web-2050
* https://mitlibraries.atlassian.net/browse/web-2052

** How does this address that need:

This adds a function to the parent theme which implements the add_filter
command provided by WordPress. This allows us to define a conditional
that assigns the cache length according to defined logic - logic which
can be pretty flexible in the checks it performs. The default response
from this function is the existing cache length of one week.

** Document any side effects to this change:

* The function is being defined within the parent theme, which gives it
  reach across the network only because everything on the network uses
  either the parent theme or one descended from it. Content within a
  site other than the parent site is a little awkward to identify, so
  there is a $site variable defined that seems to use PHP's parse_url
  to extract the site slug reliably.

* Because we are implementing this filter to occasionally define the
  max cache lifetime, the existing admin UI that also allows site
  admins to set a default cache length will get disabled. This moves
  the cache length management firmly into the realm of engineers and
  version control, and away from site admins.

* Code Climate is flagging the next bit of functions.php for a poorly
  formatted function comment - so I'm fixing it here, even though it
  isn't strictly necessary.