Extend composer multidev prompting by one step
** Why are these changes being introduced:

We define some custom Composer commands to remind engineers of the
syntax to use for multidev management. One of those commands is focused
on the search-replace commands to manipulate database contents, but this
workflow has recently found a second step to more accurately manage db
contents.

Specificially, after the blunt tool to update domain names, there needs
to be a second, more targeted, command to revert some email addresses
that need to preserved with the original value in order for WordPress to
still be able to send emails.

** Relevant ticket(s):

n/a

** How does this address that need:

This updates the existing multidev-search-replace-syntax command to also
provide the suggested syntax for this second command.

Like before, this command is only echoed to the terminal, and not run
directly.

** Document any side effects to this change:

* The search-replace command does not currently adapt to multidevs that
  have a custom domain assigned (live, test, dev, and stage). That is
  a known issue with this command, which I've chosen not to try and
  solve here.