Define theme setting for Google Maps API key
** Why are these changes being introduced:

* We received a bug report for having the Google Maps API key in our
repository as a security vulnerability.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/pw-172

** How does this address that need:

* While we disagree that this represents a security vulnerability - the
key is sent to all website users in order to load map assets - there is
also no need to handle the key directly in our source code. This defines
a theme settings field that will store this key going forward, which
will also allow us to rotate the key more easily without a code change.

* The access granted to our maps key has also been restricted, which
results in less ability for users to spot the key (as happened here)
and then probe its ability for related APIs that are not needed but
were inadvertently authorized.

** Document any side effects to this change:

* It is marginally easier for non-developers to manage this key now.